Tweaks and fixes
Fixed an issue where players could move indefinately as ling as they where cahrging the bazooka.

Some tweaks to player health and stuff

+++ Assets/Scripts/TurnsManager.cs
@@ -174,9 +174,9 @@ public class TurnsManager : MonoBehaviour
 
     }
 
-    public float GetRemainingTurnTime()
+    public int GetRemainingTurnTime()
     {
-        return _timer;
+        return ((int)_timer);
     }
 
     public GameObject GetActiveWorm()
